+++ src/SoCreate.ServiceFabric.PubSub/BrokerServiceBase.cs
@@ -262,12 +262,7 @@ namespace SoCreate.ServiceFabric.PubSub
                 var timeoutCancellationToken = linkedTokenSource.Token;
                 try
                 {
-                    await Task.WhenAll(
-                        from subscription in _queues
-                        let queueName = subscription.Key
-                        let subscriber = subscription.Value
-                        where subscriber.ShouldProcessMessages()
-                        select ProcessQueues(timeoutCancellationToken, subscriber, queueName));
+                    await ProcessAllQueuesAsync(timeoutCancellationToken);
                 }
                 catch (TaskCanceledException)
                 {//swallow and move on..
@@ -290,6 +285,16 @@ namespace SoCreate.ServiceFabric.PubSub
                 await Task.Delay(Period, cancellationToken);
             }
             // ReSharper disable once FunctionNeverReturns
+        }
+
+        protected async Task ProcessAllQueuesAsync(CancellationToken timeoutCancellationToken)
+        {
+            await Task.WhenAll(
+                from subscription in _queues
+                let queueName = subscription.Key
+                let subscriber = subscription.Value
+                where subscriber.ShouldProcessMessages()
+                select ProcessQueues(timeoutCancellationToken, subscriber, queueName));
         }
 
         /// <inheritdoc />